requirements.txt
h5py==3.8.0
imageio==2.26.0
matplotlib==3.7.1
numpy<1.24
probscale==0.2.5
pygame==2.2.0
pytest==7.2.2
pyyaml==5.4.1
sacred==0.8.4
scipy==1.10.1
seaborn==0.12.2
snakeviz==2.1.1
tensorboard==2.12.0
tensorboard-logger==0.1.0
tensorboardx==2.6
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.65.0
gym==0.21.0
wandb==0.13.11
